+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/storage/portablegrid/PortableGridItemExtendedMenuProvider.java
@@ -20,15 +20,21 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 
 class PortableGridItemExtendedMenuProvider implements ExtendedMenuProvider<PortableGridData> {
+    @Nullable
+    private final Component name;
     private final Grid grid;
     private final EnergyStorage energyStorage;
     private final DiskInventory diskInventory;
     private final SlotReference slotReference;
 
-    PortableGridItemExtendedMenuProvider(final Grid grid,
-                                         final EnergyStorage energyStorage,
-                                         final DiskInventory diskInventory,
-                                         final SlotReference slotReference) {
+    PortableGridItemExtendedMenuProvider(
+        @Nullable final Component name,
+        final Grid grid,
+        final EnergyStorage energyStorage,
+        final DiskInventory diskInventory,
+        final SlotReference slotReference
+    ) {
+        this.name = name;
         this.grid = grid;
         this.energyStorage = energyStorage;
         this.diskInventory = diskInventory;
@@ -52,7 +58,7 @@ class PortableGridItemExtendedMenuProvider implements ExtendedMenuProvider<Porta
 
     @Override
     public Component getDisplayName() {
-        return ContentNames.PORTABLE_GRID;
+        return name == null ? ContentNames.PORTABLE_GRID : name;
     }
 
     @Nullable