+++ refinedstorage-neoforge/src/main/java/com/refinedmods/refinedstorage/neoforge/datagen/loot/BlockDropProvider.java
@@ -8,9 +8,11 @@ import java.util.List;
 import java.util.Set;
 
 import net.minecraft.core.HolderLookup;
+import net.minecraft.core.component.DataComponents;
 import net.minecraft.data.loot.BlockLootSubProvider;
 import net.minecraft.world.flag.FeatureFlags;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.storage.loot.functions.CopyComponentsFunction;
 
 public class BlockDropProvider extends BlockLootSubProvider {
     public BlockDropProvider(final HolderLookup.Provider provider) {
@@ -19,24 +21,36 @@ public class BlockDropProvider extends BlockLootSubProvider {
 
     @Override
     protected void generate() {
-        Blocks.INSTANCE.getCable().forEach((color, id, block) -> dropSelf(block.get()));
-        Blocks.INSTANCE.getGrid().forEach((color, id, block) -> dropSelf(block.get()));
-        Blocks.INSTANCE.getCraftingGrid().forEach((color, id, block) -> dropSelf(block.get()));
-        Blocks.INSTANCE.getPatternGrid().forEach((color, id, block) -> dropSelf(block.get()));
+        Blocks.INSTANCE.getCable().forEach((color, id, block) -> drop(block.get()));
+        Blocks.INSTANCE.getGrid().forEach((color, id, block) -> drop(block.get()));
+        Blocks.INSTANCE.getCraftingGrid().forEach((color, id, block) -> drop(block.get()));
+        Blocks.INSTANCE.getPatternGrid().forEach((color, id, block) -> drop(block.get()));
         Blocks.INSTANCE.getController().forEach((color, id, block) -> add(
             block.get(),
-            createSingleItemTable(block.get()).apply(EnergyLootItemFunction::new)
+            createSingleItemTable(block.get())
+                .apply(EnergyLootItemFunction::new)
+                .apply(copyName())
         ));
-        Blocks.INSTANCE.getCreativeController().forEach((color, id, block) -> dropSelf(block.get()));
-        Blocks.INSTANCE.getDetector().forEach((color, id, block) -> dropSelf(block.get()));
-        Blocks.INSTANCE.getConstructor().forEach((color, id, block) -> dropSelf(block.get()));
-        Blocks.INSTANCE.getDestructor().forEach((color, id, block) -> dropSelf(block.get()));
-        Blocks.INSTANCE.getWirelessTransmitter().forEach((color, id, block) -> dropSelf(block.get()));
-        Blocks.INSTANCE.getNetworkReceiver().forEach((color, id, block) -> dropSelf(block.get()));
-        Blocks.INSTANCE.getNetworkTransmitter().forEach((color, id, block) -> dropSelf(block.get()));
-        Blocks.INSTANCE.getSecurityManager().forEach((color, id, block) -> dropSelf(block.get()));
-        Blocks.INSTANCE.getRelay().forEach((color, id, block) -> dropSelf(block.get()));
-        Blocks.INSTANCE.getDiskInterface().forEach((color, id, block) -> dropSelf(block.get()));
+        Blocks.INSTANCE.getCreativeController().forEach((color, id, block) -> drop(block.get()));
+        Blocks.INSTANCE.getDetector().forEach((color, id, block) -> drop(block.get()));
+        Blocks.INSTANCE.getConstructor().forEach((color, id, block) -> drop(block.get()));
+        Blocks.INSTANCE.getDestructor().forEach((color, id, block) -> drop(block.get()));
+        Blocks.INSTANCE.getWirelessTransmitter().forEach((color, id, block) -> drop(block.get()));
+        Blocks.INSTANCE.getNetworkReceiver().forEach((color, id, block) -> drop(block.get()));
+        Blocks.INSTANCE.getNetworkTransmitter().forEach((color, id, block) -> drop(block.get()));
+        Blocks.INSTANCE.getSecurityManager().forEach((color, id, block) -> drop(block.get()));
+        Blocks.INSTANCE.getRelay().forEach((color, id, block) -> drop(block.get()));
+        Blocks.INSTANCE.getDiskInterface().forEach((color, id, block) -> drop(block.get()));
+    }
+
+    private void drop(final Block block) {
+        add(block, createSingleItemTable(block)
+            .apply(copyName()));
+    }
+
+    private static CopyComponentsFunction.Builder copyName() {
+        return CopyComponentsFunction.copyComponents(CopyComponentsFunction.Source.BLOCK_ENTITY)
+            .include(DataComponents.CUSTOM_NAME);
     }
 
     @Override